pyproject.toml
[project]
name = "nova-skill"
version = "0.1.0"
description = "轻量级 Skill + Tool 框架"
requires-python = ">=3.10"

dependencies = [
    # LangChain ecosystem - latest stable
    "langchain>=1.2.0",
    "langchain-core>=1.2.0",
    "langchain-openai>=1.1.0",
    "langchain-anthropic>=1.3.0",
    "langgraph>=1.1.0",
    # FastAPI
    "fastapi>=0.135.0",
    "uvicorn[standard]>=0.41.0",
    "pydantic>=2.12.0",
    # Tools
    "httpx>=0.28.0",
    "pyyaml>=6.0.0",
    # Logging
    "loguru>=0.7.0",
    # Environment variables
    "python-dotenv>=1.0.0",
    "ddgs>=9.0.0",
    "tzdata>=2025.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
python-preference = "only-system"

[tool.ruff]
line-length = 100
target-version = "py310"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.26.0",
]
